predict support, evaluator cleanup
