+++ ocpmodels/trainers/ocp_trainer.py
@@ -19,7 +19,7 @@ from tqdm import tqdm
 from ocpmodels.common import distutils
 from ocpmodels.common.registry import registry
 from ocpmodels.common.relaxation.ml_relaxation import ml_relax
-from ocpmodels.common.utils import change_mat, check_traj_files, irreps_sum
+from ocpmodels.common.utils import cg_decomp_mat, check_traj_files, irreps_sum
 from ocpmodels.modules.evaluator import Evaluator
 from ocpmodels.modules.normalizer import Normalizer
 from ocpmodels.modules.scaling.util import ensure_fitted
@@ -84,6 +84,7 @@ class OCPTrainer(BaseTrainer):
         cpu=False,
         slurm={},
         noddp=False,
+        name="ocp",
     ):
         super().__init__(
             task=task,
@@ -104,503 +105,8 @@ class OCPTrainer(BaseTrainer):
             cpu=cpu,
             slurm=slurm,
             noddp=noddp,
+            name=name,
         )
-
-    def load_task(self):
-        self.targets = self.config["task"]["targets"]
-        self.num_targets = 1
-
-        self.train_targets = {}
-        for target in self.targets:
-            if "irreps" in self.targets[target]:
-                self.train_targets[target] = self.targets[target]
-            else:
-                for subtarget in self.targets[target]:
-                    self.train_targets[subtarget] = self.targets[target][
-                        subtarget
-                    ]
-                    self.train_targets[subtarget]["parent"] = target
-
-        # Normalizer for the dataset.
-        self.normalizers = {}
-        for target in self.train_targets:
-            self.normalizers[target] = Normalizer(
-                mean=self.train_targets.get("mean", 0),
-                std=self.train_targets.get("std", 1),
-                device=self.device,
-            )
-
-        self.eval_metrics = self.config["task"]["metrics"]
-
-        # assert len(self.targets.keys() - self.eval_metrics.keys()) == 0
-
-    # Takes in a new data source and generates predictions on it.
-    @torch.no_grad()
-    def predict(
-        self,
-        data_loader,
-        per_image=True,
-        results_file=None,
-        disable_tqdm=False,
-    ):
-        ensure_fitted(self._unwrapped_model, warn=True)
-
-        if distutils.is_master() and not disable_tqdm:
-            logging.info("Predicting on test.")
-        assert isinstance(
-            data_loader,
-            (
-                torch.utils.data.dataloader.DataLoader,
-                torch_geometric.data.Batch,
-            ),
-        )
-        rank = distutils.get_rank()
-
-        if isinstance(data_loader, torch_geometric.data.Batch):
-            data_loader = [[data_loader]]
-
-        self.model.eval()
-        if self.ema:
-            self.ema.store()
-            self.ema.copy_to()
-
-        if self.normalizers is not None and "target" in self.normalizers:
-            self.normalizers["target"].to(self.device)
-            self.normalizers["grad_target"].to(self.device)
-
-        predictions = {"id": [], "energy": [], "forces": [], "chunk_idx": []}
-
-        for i, batch_list in tqdm(
-            enumerate(data_loader),
-            total=len(data_loader),
-            position=rank,
-            desc="device {}".format(rank),
-            disable=disable_tqdm,
-        ):
-            with torch.cuda.amp.autocast(enabled=self.scaler is not None):
-                out = self._forward(batch_list)
-
-            if self.normalizers is not None and "target" in self.normalizers:
-                out["energy"] = self.normalizers["target"].denorm(
-                    out["energy"]
-                )
-                out["forces"] = self.normalizers["grad_target"].denorm(
-                    out["forces"]
-                )
-            if per_image:
-                systemids = [
-                    str(i) + "_" + str(j)
-                    for i, j in zip(
-                        batch_list[0].sid.tolist(), batch_list[0].fid.tolist()
-                    )
-                ]
-                predictions["id"].extend(systemids)
-                batch_natoms = torch.cat(
-                    [batch.natoms for batch in batch_list]
-                )
-                batch_fixed = torch.cat([batch.fixed for batch in batch_list])
-                # total energy target requires predictions to be saved in float32
-                # default is float16
-                if (
-                    self.config["task"].get("prediction_dtype", "float16")
-                    == "float32"
-                    or self.config["task"]["dataset"] == "oc22_lmdb"
-                ):
-                    predictions["energy"].extend(
-                        out["energy"].cpu().detach().to(torch.float32).numpy()
-                    )
-                    forces = out["forces"].cpu().detach().to(torch.float32)
-                else:
-                    predictions["energy"].extend(
-                        out["energy"].cpu().detach().to(torch.float16).numpy()
-                    )
-                    forces = out["forces"].cpu().detach().to(torch.float16)
-                per_image_forces = torch.split(forces, batch_natoms.tolist())
-                per_image_forces = [
-                    force.numpy() for force in per_image_forces
-                ]
-                # evalAI only requires forces on free atoms
-                if results_file is not None:
-                    _per_image_fixed = torch.split(
-                        batch_fixed, batch_natoms.tolist()
-                    )
-                    _per_image_free_forces = [
-                        force[(fixed == 0).tolist()]
-                        for force, fixed in zip(
-                            per_image_forces, _per_image_fixed
-                        )
-                    ]
-                    _chunk_idx = np.array(
-                        [
-                            free_force.shape[0]
-                            for free_force in _per_image_free_forces
-                        ]
-                    )
-                    per_image_forces = _per_image_free_forces
-                    predictions["chunk_idx"].extend(_chunk_idx)
-                predictions["forces"].extend(per_image_forces)
-            else:
-                predictions["energy"] = out["energy"].detach()
-                predictions["forces"] = out["forces"].detach()
-                if self.ema:
-                    self.ema.restore()
-                return predictions
-
-        predictions["forces"] = np.array(predictions["forces"])
-        predictions["chunk_idx"] = np.array(predictions["chunk_idx"])
-        predictions["energy"] = np.array(predictions["energy"])
-        predictions["id"] = np.array(predictions["id"])
-        self.save_results(
-            predictions, results_file, keys=["energy", "forces", "chunk_idx"]
-        )
-
-        if self.ema:
-            self.ema.restore()
-
-        return predictions
-
-    def update_best(
-        self,
-        primary_metric,
-        val_metrics,
-        disable_eval_tqdm=True,
-    ):
-        if (
-            "mae" in primary_metric
-            and val_metrics[primary_metric]["metric"] < self.best_val_metric
-        ) or (
-            "mae" not in primary_metric
-            and val_metrics[primary_metric]["metric"] > self.best_val_metric
-        ):
-            self.best_val_metric = val_metrics[primary_metric]["metric"]
-            self.save(
-                metrics=val_metrics,
-                checkpoint_file="best_checkpoint.pt",
-                training_state=False,
-            )
-            if self.test_loader is not None:
-                self.predict(
-                    self.test_loader,
-                    results_file="predictions",
-                    disable_tqdm=disable_eval_tqdm,
-                )
-
-    def train(self, disable_eval_tqdm=False):
-        ensure_fitted(self._unwrapped_model, warn=True)
-
-        eval_every = self.config["optim"].get(
-            "eval_every", len(self.train_loader)
-        )
-        checkpoint_every = self.config["optim"].get(
-            "checkpoint_every", eval_every
-        )
-        primary_metric = self.config["task"]["primary_metric"]
-        # TODO: support for old naming conventions - is2re, s2ef, etc.
-        # primary_metric = self.config["task"].get(
-        # "primary_metric", self.evaluator.task_primary_metric[self.name]
-        # )
-        if (
-            not hasattr(self, "primary_metric")
-            or self.primary_metric != primary_metric
-        ):
-            self.best_val_metric = 1e9 if "mae" in primary_metric else -1.0
-        else:
-            primary_metric = self.primary_metric
-        self.metrics = {}
-
-        # Calculate start_epoch from step instead of loading the epoch number
-        # to prevent inconsistencies due to different batch size in checkpoint.
-        start_epoch = self.step // len(self.train_loader)
-
-        for epoch_int in range(
-            start_epoch, self.config["optim"]["max_epochs"]
-        ):
-            self.train_sampler.set_epoch(epoch_int)
-            skip_steps = self.step % len(self.train_loader)
-            train_loader_iter = iter(self.train_loader)
-
-            for i in range(skip_steps, len(self.train_loader)):
-                self.epoch = epoch_int + (i + 1) / len(self.train_loader)
-                self.step = epoch_int * len(self.train_loader) + i + 1
-                self.model.train()
-
-                # Get a batch.
-                batch = next(train_loader_iter)
-
-                # Forward, loss, backward.
-                with torch.cuda.amp.autocast(enabled=self.scaler is not None):
-                    out = self._forward(batch)
-                    loss = self._compute_loss(out, batch)
-                loss = self.scaler.scale(loss) if self.scaler else loss
-                self._backward(loss)
-                scale = self.scaler.get_scale() if self.scaler else 1.0
-
-                # Compute metrics.
-                self.metrics = self._compute_metrics(
-                    out,
-                    batch,
-                    self.evaluator,
-                    self.metrics,
-                )
-                self.metrics = self.evaluator.update(
-                    "loss", loss.item() / scale, self.metrics
-                )
-
-                # Log metrics.
-                log_dict = {k: self.metrics[k]["metric"] for k in self.metrics}
-                log_dict.update(
-                    {
-                        "lr": self.scheduler.get_lr(),
-                        "epoch": self.epoch,
-                        "step": self.step,
-                    }
-                )
-                if (
-                    self.step % self.config["cmd"]["print_every"] == 0
-                    and distutils.is_master()
-                    and not self.is_hpo
-                ):
-                    log_str = [
-                        "{}: {:.2e}".format(k, v) for k, v in log_dict.items()
-                    ]
-                    logging.info(", ".join(log_str))
-                    self.metrics = {}
-
-                if self.logger is not None:
-                    self.logger.log(
-                        log_dict,
-                        step=self.step,
-                        split="train",
-                    )
-
-                if (
-                    checkpoint_every != -1
-                    and self.step % checkpoint_every == 0
-                ):
-                    self.save(
-                        checkpoint_file="checkpoint.pt", training_state=True
-                    )
-
-                # Evaluate on val set every `eval_every` iterations.
-                if self.step % eval_every == 0:
-                    if self.val_loader is not None:
-                        val_metrics = self.validate(
-                            split="val",
-                            disable_tqdm=disable_eval_tqdm,
-                        )
-                        self.update_best(
-                            primary_metric,
-                            val_metrics,
-                            disable_eval_tqdm=disable_eval_tqdm,
-                        )
-                        if self.is_hpo:
-                            self.hpo_update(
-                                self.epoch,
-                                self.step,
-                                self.metrics,
-                                val_metrics,
-                            )
-
-                    if self.config["task"].get("eval_relaxations", False):
-                        if "relax_dataset" not in self.config["task"]:
-                            logging.warning(
-                                "Cannot evaluate relaxations, relax_dataset not specified"
-                            )
-                        else:
-                            self.run_relaxations()
-
-                if self.scheduler.scheduler_type == "ReduceLROnPlateau":
-                    if self.step % eval_every == 0:
-                        self.scheduler.step(
-                            metrics=val_metrics[primary_metric]["metric"],
-                        )
-                else:
-                    self.scheduler.step()
-
-            torch.cuda.empty_cache()
-
-            if checkpoint_every == -1:
-                self.save(checkpoint_file="checkpoint.pt", training_state=True)
-
-        self.train_dataset.close_db()
-        if self.config.get("val_dataset", False):
-            self.val_dataset.close_db()
-        if self.config.get("test_dataset", False):
-            self.test_dataset.close_db()
-
-    def _forward(self, batch_list):
-        # forward pass.
-        return self.model(batch_list)
-
-    def _compute_loss(self, out, batch_list):
-        natoms = torch.cat(
-            [batch.natoms.to(self.device) for batch in batch_list], dim=0
-        )
-        natoms = torch.repeat_interleave(natoms, natoms)
-        batch_size = natoms.numel()
-
-        loss = []
-        if self.config["task"].get("train_on_free_atoms", True):
-            fixed = torch.cat(
-                [batch.fixed.to(self.device) for batch in batch_list]
-            )
-            mask = fixed == 0
-
-        for target_name in self.train_targets:
-            if "parent" not in self.train_targets[target_name]:
-                target = torch.cat(
-                    [
-                        batch[target_name].to(self.device)
-                        for batch in batch_list
-                    ],
-                    dim=0,
-                )
-            # property is a decomposition of a higher order tensor
-            else:
-                irreps = self.train_targets[target_name]["irreps"]
-                if irreps > 2:
-                    raise NotImplementedError
-
-                target = [
-                    torch.einsum(
-                        "ab, cb->ca",
-                        change_mat.to(self.device),
-                        batch[self.train_targets[target_name]["parent"]],
-                    )
-                    for batch in batch_list
-                ]
-
-                target = torch.cat(
-                    [
-                        batch[
-                            :,
-                            max(0, irreps_sum(irreps - 1)) : irreps_sum(
-                                irreps
-                            ),
-                        ]
-                        for batch in target
-                    ],
-                    dim=0,
-                )
-
-            pred = out[target_name]
-
-            if (
-                self.config["task"].get("train_on_free_atoms", True)
-                and self.train_targets[target_name].get("level", "system")
-                == "atom"
-            ):
-                target = target[mask]
-                pred = pred[mask]
-                natoms = natoms[mask]
-
-            if self.normalizers.get(target_name, False):
-                target = self.normalizers[target_name].norm(target)
-
-            mult = self.train_targets[target_name].get("coefficient", 1)
-
-            loss.append(
-                mult
-                * self.loss_fn[target_name](
-                    pred,
-                    target,
-                    natoms=natoms,
-                    batch_size=batch_size,
-                )
-            )
-
-        # Sanity check to make sure the compute graph is correct.
-        for lc in loss:
-            assert hasattr(lc, "grad_fn")
-
-        loss = sum(loss)
-        return loss
-
-    def _compute_metrics(self, out, batch_list, evaluator, metrics={}):
-        natoms = torch.cat(
-            [batch.natoms.to(self.device) for batch in batch_list], dim=0
-        )
-
-        if self.config["task"].get("eval_on_free_atoms", True):
-            fixed = torch.cat(
-                [batch.fixed.to(self.device) for batch in batch_list]
-            )
-            mask = fixed == 0
-
-            s_idx = 0
-            natoms_free = []
-            for _natoms in natoms:
-                natoms_free.append(
-                    torch.sum(mask[s_idx : s_idx + _natoms]).item()
-                )
-                s_idx += _natoms
-            natoms = torch.LongTensor(natoms_free).to(self.device)
-
-        targets = {}
-        for target_name in self.train_targets:
-            if "parent" not in self.train_targets[target_name]:
-                target = torch.cat(
-                    [
-                        batch[target_name].to(self.device)
-                        for batch in batch_list
-                    ],
-                    dim=0,
-                )
-            else:
-                irreps = self.train_targets[target_name]["irreps"]
-                parent_target_name = self.train_targets[target_name]["parent"]
-
-                if parent_target_name not in targets:
-                    parent_target = torch.cat(
-                        [
-                            batch[parent_target_name].to(self.device)
-                            for batch in batch_list
-                        ],
-                        dim=0,
-                    )
-                    targets[parent_target_name] = parent_target
-
-                target = [
-                    torch.einsum(
-                        "ab, cb->ca",
-                        change_mat.to(self.device),
-                        batch[parent_target_name],
-                    )
-                    for batch in batch_list
-                ]
-
-                target = torch.cat(
-                    [
-                        batch[
-                            :,
-                            max(0, irreps_sum(irreps - 1)) : irreps_sum(
-                                irreps
-                            ),
-                        ]
-                        for batch in target
-                    ],
-                    dim=0,
-                )
-
-            if (
-                self.config["task"].get("eval_on_free_atoms", True)
-                and self.train_targets[target_name].get("level", "system")
-                == "atom"
-            ):
-                target = target[mask]
-                out[target_name] = out[target_name][mask]
-
-            targets[target_name] = target
-            if self.normalizers.get(target_name, False):
-                out[target_name] = self.normalizers[target_name].denorm(
-                    out[target_name]
-                )
-
-        targets["natoms"] = natoms
-        out["natoms"] = natoms
-
-        metrics = evaluator.eval(out, targets, prev_metrics=metrics)
-        return metrics
 
     def run_relaxations(self, split="val"):
         ensure_fitted(self._unwrapped_model)
